fix: Disable rich interactivity when running in CI
