+++ src/pip/_internal/utils/logging.py
@@ -29,7 +29,7 @@ from pip._vendor.rich.style import Style
 from pip._internal.utils._log import VERBOSE, getLogger
 from pip._internal.utils.compat import WINDOWS
 from pip._internal.utils.deprecation import DEPRECATION_MSG_PREFIX
-from pip._internal.utils.misc import StreamWrapper, ensure_dir
+from pip._internal.utils.misc import StreamWrapper, ensure_dir, looks_like_ci
 
 _log_state = threading.local()
 _stdout_console = None
@@ -320,8 +320,15 @@ def setup_logging(verbosity: int, no_color: bool, user_log_file: str | None) -> 
         ["user_log"] if include_user_log else []
     )
     global _stdout_console, stderr_console
-    _stdout_console = PipConsole(file=sys.stdout, no_color=no_color, soft_wrap=True)
-    _stderr_console = PipConsole(file=sys.stderr, no_color=no_color, soft_wrap=True)
+    console_options: dict[str, Any] = {"no_color": no_color, "soft_wrap": True}
+    if looks_like_ci():
+        # A lot of CI workflows use FORCE_COLOR and similar environment
+        # variables so their CI output remains colorized, but this has
+        # the unfortunate consequence of rendering progress bars that
+        # (often) span many lines.
+        console_options["force_interactive"] = False
+    _stdout_console = PipConsole(file=sys.stdout, **console_options)
+    _stderr_console = PipConsole(file=sys.stderr, **console_options)
 
     logging.config.dictConfig(
         {
